% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecov.R
\name{codecov}
\alias{codecov}
\title{Upload test coverage results to codecov}
\usage{
codecov(
  path = ".",
  coverage = NULL,
  url = NULL,
  service = NULL,
  token = NULL,
  name = NULL,
  commit = NULL,
  branch = NULL,
  build = NULL,
  job = NULL,
  build_url = NULL,
  slug = NULL,
  yaml = NULL,
  flags = NULL,
  pr = NULL
)
}
\arguments{
\item{path}{Path to the package tree.}

\item{coverage}{Test coverage results. If \code{NULL} then the last
results are used via \code{\link[=last]{last()}}.}

\item{url}{Codecov URL. If \code{NULL} the default URL is used.}

\item{service}{The CI service name. If \code{NULL} then it is detected
automatically from environment variables.}

\item{token}{Codecode token to identify the project. If the
\code{CODECOV_TOKEN} environment variable is set, it is used as the token.}

\item{name}{A custom name for this specific upload.}

\item{commit}{The destination commit sha for the report. If \code{NULL} it is
detected automatically.}

\item{branch}{The target branch for the report. If \code{NULL} it is
detected automatically.}

\item{build}{The build number provided by your CI service. If \code{NULL} it is
detected automatically.}

\item{job}{The job number provided by your CI service. If \code{NULL} it is
detected automatically.}

\item{build_url}{The http url to link back to your CI provider. If \code{NULL}
it is detected automatically.}

\item{slug}{The \verb{owner/repo`` slug name of the project. If }NULL` it is
detected automatically.}

\item{yaml}{The relative path to the \verb{codecov.yml`` in this project. If }NULL` it is detected automatically.}

\item{flags}{Used for \href{https://docs.codecov.com/v4.6/docs/flags}{flags}.
Can be one or more flags. E.g., \code{flags = "unit"} or \code{flags = "unit,java"}.}

\item{pr}{The pull request number this commit is currently found in.
If \code{NULL} it is detected automatically.}
}
\description{
Upload test coverage results to codecov
}
\details{
Currently supported CI services:
\itemize{
\item GitHub Actions.
}
}
