% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage-report.R
\name{coverage_report_file}
\alias{coverage_report_file}
\title{Generate a HTML coverage report for a source file}
\usage{
coverage_report_file(
  code_file = NULL,
  coverage = NULL,
  path = ".",
  output_dir = NULL,
  show = interactive()
)
}
\arguments{
\item{code_file}{Relative path to the code file to show.}

\item{coverage}{A coverage object, optionally.
If not present, the output of \code{\link[=last_coverage_results]{last_coverage_results()}} is used.}

\item{path}{Path to the package tree.}

\item{output_dir}{Output directory to put the HTML report into.}

\item{show}{Whether to open a browser window to show the HTML
report.}
}
\description{
\code{\link[=coverage_report]{coverage_report()}} calls this function for all source files, so
you only need this if you want a report for a single file.
}
