% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-coverage.R
\name{diff}
\alias{diff}
\title{Show test coverage for new code, relative to a git branch or commit}
\usage{
diff(
  mode = c("auto", "diff", "branch", "commit", "pr"),
  path = ".",
  coverage = NULL
)
}
\arguments{
\item{mode}{The type of comparison to perform to determine the new
code that is analyzed. Possible values:
\itemize{
\item \code{auto}: will choose \code{branch} if \code{path} is on the non-default (\code{main}
or \code{master}) branch, and \code{diff} otherwise.
\item \code{diff}: compare the current code to \code{HEAD}.
\item \code{branch}: compare the current code to the default branch, \code{main} or
\code{master}.
\item \code{commit}: analyze the last commit, i.e. compare \code{HEAD} to \code{HEAD~1}.
\item \code{pr}: analyze the current pull request. This only works if called
on Github Actions.
}}

\item{path}{Path to the package tree.}

\item{coverage}{Test coverage results. If \code{NULL} then the last
results are used via \code{\link[=last]{last()}}.}
}
\description{
Show test coverage for new code, relative to a git branch or commit
}
