% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-dev.R
\name{document}
\alias{document}
\alias{d}
\title{Run roxygen2 to generate the package manual and namespace files}
\usage{
document(path = ".", clean = FALSE, local_install = TRUE)

d(path = ".", clean = FALSE, local_install = TRUE)
}
\arguments{
\item{path}{Path to the package tree.}

\item{clean}{Whether to delete the build directory before the build.}

\item{local_install}{Whether to install the built package into a local
library and add that library to the beginning of the library path.}
}
\value{
The same as \code{\link[=reload]{reload()}}, with an additional field \code{roxy},
that contains the return value of \code{\link[roxygen2:roxygenize]{roxygen2::roxygenize()}}.
}
\description{
Loads the debug build, and calls \code{\link[roxygen2:roxygenize]{roxygen2::roxygenize()}}.
}
\details{
\code{d()} is an alias of \code{document()}.
}
