% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install local package tree}
\usage{
install(
  type = c("release", "debug", "coverage"),
  path = ".",
  makeflags = NULL,
  clean = FALSE,
  lib = NULL,
  INSTALL_opts = c("--no-staged-install", "--no-test-load"),
  ...
)
}
\arguments{
\item{type}{Build type. One of:
\itemize{
\item \code{"debug"}: Compiled with debug flags. (Passes \code{debug = TRUE} to
\code{\link[pkgbuild:compiler_flags]{pkgbuild::compiler_flags()}}.)
\item \code{"release"}: Compiled without debug flags. (Passes \code{debug = FALSE} to
\code{\link[pkgbuild:compiler_flags]{pkgbuild::compiler_flags()}}.)
\item \code{"coverage"}: Compiled with debug flags and code coverage support.
}}

\item{path}{Path to the package tree.}

\item{makeflags}{Named character vector to override the default make flags.}

\item{clean}{Whether to delete the build directory before the build.}

\item{lib, INSTALL_opts, ...}{Additional arguments are passed to
\code{\link[utils:install.packages]{utils::install.packages()}}.}
}
\description{
Install local package tree
}
