% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{last}
\alias{last}
\title{Find and print the last test coverage results}
\usage{
last(path = ".")
}
\arguments{
\item{path}{Path to the package tree.}
}
\value{
A package_coverage object. If there are no previous results,
then a message is shown and \code{NULL} is returned.
}
\description{
Find and print the last test coverage results
}
