% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-dev.R
\name{last_coverage_results}
\alias{last_coverage_results}
\title{Find and print the last test coverage results}
\usage{
last_coverage_results(path = ".")
}
\arguments{
\item{path}{Path to the package tree.}
}
\value{
A package_coverage object. If there are no previous results,
  then a message is shown and `NULL` is returned.
}
\description{
Find and print the last test coverage results
}
