% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-dev.R
\name{last_test_results}
\alias{last_test_results}
\title{Find and print the last test results}
\usage{
last_test_results(path = ".")
}
\arguments{
\item{path}{Path to the package tree.}
}
\value{
The test results in a list with class testthat_results. If there
  are no previous results, then a message is shown and `NULL` is
  returned.
}
\description{
Find and print the last test results
}
