% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-dev.R
\name{list_builds}
\alias{list_builds}
\title{List development builds}
\usage{
list_builds(path = ".")
}
\arguments{
\item{path}{Path to the package tree.}
}
\value{
Data frame with columns:
  - `type`: Build type, character.
  - `r_version`: R version, first two digits only, character.
  - `platform`: Build platform triplet, character.
  - `last_built`: Time stamp of the last build, [POSIXct].
  - `dist_size`: Size of the build directory in bytes, double.
  - `id`: Name of the build directory.
}
\description{
List development builds
}
