% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-dev.R
\name{load_package}
\alias{load_package}
\alias{l}
\title{Load a package tree for development}
\usage{
load_package(
  type = c("debug", "release", "coverage"),
  path = ".",
  makeflags = NULL,
  clean = FALSE,
  local_install = TRUE
)

l(
  type = c("debug", "release", "coverage"),
  path = ".",
  makeflags = NULL,
  clean = FALSE,
  local_install = TRUE
)
}
\arguments{
\item{type}{Build type. One of:
- `"debug"`: Compiled with debug flags. (Passes `debug = TRUE` to
  [pkgbuild::compiler_flags()].)
- `"release"`: Compiled without debug flags. (Passes `debug = FALSE` to
  [pkgbuild::compiler_flags()].)
- `"coverage"`: Compiled with debug flags and code coverage support.}

\item{path}{Path to the package tree.}

\item{makeflags}{Named character vector to override the default make flags.}

\item{clean}{Whether to delete the build directory before the build.}

\item{local_install}{Whether to install the built package into a local
library and add that library to the beginning of the library path.}
}
\value{
A list with entries:
- `setup`: List of the build setup. Entries:
  - `version`: Version of format of the build directory.
  - `rver`: R version. (First two digits only.)
  - `platform`: Build platform triplet.
  - `type`: The build type.
  - `hash`: The hash of the setup that is used as the build directory
    name.
  - `compiler_flags`: Nemed character vector of extra `Makevars` flags
    to use.
  - `dir`: The build directory.
  - `pkgname`: The name of the package.
- `plan`: Data frame of the plan to create the build directory. Columns:
  - `path`: Relative path of the file.
  - `isdir`: Whether it is a directory.
  - `action`: How to create the path in the build directory: `"link"` or
    `"copy"`.
  - `target`: Relative path of the file or directory in the build
    directory.
  - `hash`: Hash of the file for files, it is `NA` for directories.
- `load`: The return value of [pkgload::load_all()].
- `coverage`: For `type = "coverage"` builds a data frame with the
  code coverage instrumentation data for R files. See
  [test_package()] for the structure.
}
\description{
A wrapper on [pkgload::load_all()] to load a package tree from a
dev build.
}
\details{
`l()` is an alias of `load_package()`.
}
