% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{markdown}
\alias{markdown}
\title{Generate a markdown test coverage report}
\usage{
markdown(path = ".", coverage = NULL, output = NULL, truncate = TRUE)
}
\arguments{
\item{path}{Path to the package tree.}

\item{coverage}{Test coverage results. If \code{NULL} then the last
results are used via \code{\link[=last]{last()}}.}

\item{output}{Output file. By default it is placed in the dev directory.}

\item{truncate}{Whether to truncate the file if it is longer than the
GitHub Actions limit (currently 1.048576 × 10\if{html}{\out{<sup>}}6\if{html}{\out{</sup>}} bytes).}
}
\value{
The path of the output file, invisibly.
}
\description{
This is to create a \href{https://github.com/features/actions}{GitHub Actions}
\href{https://docs.github.com/en/actions/reference/workflows-and-actions/workflow-commands#adding-a-job-summary}{job summary}.
}
