% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload.R
\name{reload}
\alias{reload}
\title{Load a package tree for development}
\usage{
reload(
  type = c("debug", "release", "coverage"),
  path = ".",
  makeflags = NULL,
  clean = FALSE,
  local_install = TRUE
)
}
\arguments{
\item{type}{Build type. One of:
\itemize{
\item \code{"debug"}: Compiled with debug flags. (Passes \code{debug = TRUE} to
\code{\link[pkgbuild:compiler_flags]{pkgbuild::compiler_flags()}}.)
\item \code{"release"}: Compiled without debug flags. (Passes \code{debug = FALSE} to
\code{\link[pkgbuild:compiler_flags]{pkgbuild::compiler_flags()}}.)
\item \code{"coverage"}: Compiled with debug flags and code coverage support.
}}

\item{path}{Path to the package tree.}

\item{makeflags}{Named character vector to override the default make flags.}

\item{clean}{Whether to delete the build directory before the build.}

\item{local_install}{Whether to install the built package into a local
library and add that library to the beginning of the library path.}
}
\value{
A list with entries:
\itemize{
\item \code{setup}: List of the build setup. Entries:
\itemize{
\item \code{version}: Version of format of the build directory.
\item \code{rver}: R version. (First two digits only.)
\item \code{platform}: Build platform triplet.
\item \code{type}: The build type.
\item \code{hash}: The hash of the setup that is used as the build directory
name.
\item \code{compiler_flags}: Nemed character vector of extra \code{Makevars} flags
to use.
\item \code{dir}: The build directory.
\item \code{pkgname}: The name of the package.
}
\item \code{plan}: Data frame of the plan to create the build directory. Columns:
\itemize{
\item \code{path}: Relative path of the file.
\item \code{isdir}: Whether it is a directory.
\item \code{target}: Relative path of the file or directory in the build
directory.
\item \code{hash}: Hash of the file for files, it is \code{NA} for directories.
}
\item \code{load}: The return value of \code{\link[pkgload:load_all]{pkgload::load_all()}}.
\item \code{coverage}: For \code{type = "coverage"} builds a data frame with the
code coverage instrumentation data for R files. See
\code{\link[=test]{test()}} for the structure.
}
}
\description{
A wrapper on \code{\link[pkgload:load_all]{pkgload::load_all()}} to load a package tree from a
dev build.
}
