% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage-report.R
\name{report}
\alias{report}
\title{Generate a HTML coverage report}
\usage{
report(path = ".", coverage = NULL, output = NULL, show = interactive())
}
\arguments{
\item{path}{Path to the package tree.}

\item{coverage}{Test coverage results. If \code{NULL} then the last
results are used via \code{\link[=last]{last()}}.}

\item{output}{Output file. This is the HTML index file. The rest of
of the HTML files for the source code files will be placed next
to it. By default it is generated in the dev directory.}

\item{show}{Whether to open a browser window to show the HTML
report.}
}
\value{
The path of the output file, invisibly.
}
\description{
Generate a HTML coverage report
}
