% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-dev.R
\name{rerun_failing_tests}
\alias{rerun_failing_tests}
\title{Re-run the test files that had failing tests in the last test run}
\usage{
rerun_failing_tests(
  path = ".",
  types = c("fail", "warning", "skip", "all")[1],
  show_coverage = FALSE,
  ...
)
}
\arguments{
\item{path}{Path to the package tree.}

\item{types}{Test result types to re-run, a character vector, possible
elements are `"fail"` (default), `"warning"`, `"skip"`, "`all`".
`"all"` is equivalent to `c("fail", "warning", "skip")`.}

\item{show_coverage}{Whether to show code coverage results.}

\item{...}{Additional arguments are passed to [test_package()].}
}
\description{
Re-run the test files that had failing tests in the last test run
}
