% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-dev.R
\name{retest}
\alias{retest}
\title{Re-run the test files that had failing tests in the last test run}
\usage{
retest(
  path = ".",
  types = c("fail", "warning", "skip", "all")[1],
  show_coverage = FALSE,
  ...
)
}
\arguments{
\item{path}{Path to the package tree.}

\item{types}{Test result types to re-run, a character vector, possible
elements are \code{"fail"} (default), \code{"warning"}, \code{"skip"}, "\code{all}".
\code{"all"} is equivalent to \code{c("fail", "warning", "skip")}.}

\item{show_coverage}{Whether to show code coverage results.}

\item{...}{Additional arguments are passed to \code{\link[=test]{test()}}.}
}
\description{
Re-run the test files that had failing tests in the last test run
}
