% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-active.R
\name{test_active}
\alias{test_active}
\title{Test the active file}
\usage{
test_active(file = NULL, ...)
}
\arguments{
\item{file}{Use this file instead of auto-detecting the active file.}

\item{...}{Additional arguments are passed to \code{\link[=test]{test()}}.}
}
\value{
The same as \code{\link[=test]{test()}}.
}
\description{
Calls \code{\link[=test]{test()}} with a \code{filter} to only run tests and show test coverage
for the currently active file in RStudio, Positron or VS Code.
}
