% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthatlabs-package.R
\docType{package}
\name{testthatlabs-package}
\alias{testthatlabs}
\alias{testthatlabs-package}
\title{testthatlabs: Experimental add-ons for testthat}
\description{
Experimental extensions for testthat.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gaborcsardi/testthatlabs}
  \item Report bugs at \url{https://github.com/gaborcsardi/testthatlabs/issues}
}

}
\keyword{internal}
