% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncovr-package.R
\docType{package}
\name{uncovr-package}
\alias{uncovr}
\alias{uncovr-package}
\title{uncovr: Experimental add-ons for testthat}
\description{
Experimental extensions for testthat.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gaborcsardi/uncovr}
  \item Report bugs at \url{https://github.com/gaborcsardi/uncovr/issues}
}

}
\keyword{internal}
