% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcov.R
\name{write_lconv_info}
\alias{write_lconv_info}
\title{Write an LCOV coverage report}
\usage{
write_lconv_info(path = ".", coverage = NULL, output = NULL)
}
\arguments{
\item{path}{Path to the package tree.}

\item{coverage}{Test coverage results. If `NULL` then the last
results are used via [last_coverage_results()].}

\item{output}{Path of the output file. If `NULL`, then it is
`lcov.info` inside the build directory.}
}
\value{
The path of the output file, invisibly.
}
\description{
An LCOV coverage report contains information about the code coverage
of all files in a project.
}
