% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcov.R
\name{write_lcov_info}
\alias{write_lcov_info}
\title{Write an LCOV coverage report}
\usage{
write_lcov_info(path = ".", coverage = NULL, output = NULL)
}
\arguments{
\item{path}{Path to the package tree.}

\item{coverage}{Test coverage results. If \code{NULL} then the last
results are used via \code{\link[=last_coverage_results]{last_coverage_results()}}.}

\item{output}{Path of the output file. If \code{NULL}, then it is
\verb{_lcov/lcov.info} inside the build directory.}
}
\value{
The path of the output file, invisibly.
}
\description{
An LCOV coverage report contains information about the code coverage
of all files in a project.
}
