% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-dev.R
\name{install_package}
\alias{install_package}
\alias{i}
\title{Install local package tree}
\usage{
install_package(
  type = c("release", "debug", "coverage"),
  path = ".",
  makeflags = NULL,
  clean = FALSE,
  lib = NULL,
  INSTALL_opts = c("--no-staged-install", "--no-test-load"),
  ...
)

i(
  type = c("release", "debug", "coverage"),
  path = ".",
  makeflags = NULL,
  clean = FALSE,
  lib = NULL,
  INSTALL_opts = c("--no-staged-install", "--no-test-load"),
  ...
)
}
\arguments{
\item{type}{Build type. One of:
- `"debug"`: Compiled with debug flags. (Passes `debug = TRUE` to
  [pkgbuild::compiler_flags()].)
- `"release"`: Compiled without debug flags. (Passes `debug = FALSE` to
  [pkgbuild::compiler_flags()].)
- `"coverage"`: Compiled with debug flags and code coverage support.}

\item{path}{Path to the package tree.}

\item{makeflags}{Named character vector to override the default make flags.}

\item{clean}{Whether to delete the build directory before the build.}

\item{INSTALL_opts, ...}{Additional arguments are passed to [utils::install.packages()].}
}
\description{
Install local package tree
}
\details{
`i()` is an alias of `install_package()`.
}
