% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-dev.R
\name{list_builds}
\alias{list_builds}
\title{List development builds}
\usage{
list_builds(path = ".")
}
\arguments{
\item{path}{Path to the package tree.}
}
\value{
Data frame with columns:
\itemize{
\item \code{type}: Build type, character.
\item \code{r_version}: R version, first two digits only, character.
\item \code{platform}: Build platform triplet, character.
\item \code{last_built}: Time stamp of the last build, \link{POSIXct}.
\item \code{dist_size}: Size of the build directory in bytes, double.
\item \code{id}: Name of the build directory.
}
}
\description{
List development builds
}
